rename safety/safety/safety_* to safety/modes/*
