+++ opendbc/safety/modes/hyundai.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "opendbc/safety/safety_declarations.h"
-#include "opendbc/safety/safety/safety_hyundai_common.h"
+#include "opendbc/safety/modes/hyundai_common.h"
 
 #define HYUNDAI_LIMITS(steer, rate_up, rate_down) { \
   .max_torque = (steer), \
