+++ opendbc/safety/modes/hyundai_common.h
