+++ opendbc/safety/modes/rivian.h
