+++ opendbc/safety/modes/subaru_preglobal.h
