+++ opendbc/safety/safety.h
@@ -4,27 +4,27 @@
 #include "opendbc/safety/board/can.h"
 
 // include the safety policies.
-#include "opendbc/safety/safety/safety_defaults.h"
-#include "opendbc/safety/safety/safety_honda.h"
-#include "opendbc/safety/safety/safety_toyota.h"
-#include "opendbc/safety/safety/safety_tesla.h"
-#include "opendbc/safety/safety/safety_gm.h"
-#include "opendbc/safety/safety/safety_ford.h"
-#include "opendbc/safety/safety/safety_hyundai.h"
-#include "opendbc/safety/safety/safety_chrysler.h"
-#include "opendbc/safety/safety/safety_rivian.h"
-#include "opendbc/safety/safety/safety_subaru.h"
-#include "opendbc/safety/safety/safety_subaru_preglobal.h"
-#include "opendbc/safety/safety/safety_mazda.h"
-#include "opendbc/safety/safety/safety_nissan.h"
-#include "opendbc/safety/safety/safety_volkswagen_mqb.h"
-#include "opendbc/safety/safety/safety_volkswagen_pq.h"
-#include "opendbc/safety/safety/safety_elm327.h"
-#include "opendbc/safety/safety/safety_body.h"
+#include "opendbc/safety/modes/defaults.h"
+#include "opendbc/safety/modes/honda.h"
+#include "opendbc/safety/modes/toyota.h"
+#include "opendbc/safety/modes/tesla.h"
+#include "opendbc/safety/modes/gm.h"
+#include "opendbc/safety/modes/ford.h"
+#include "opendbc/safety/modes/hyundai.h"
+#include "opendbc/safety/modes/chrysler.h"
+#include "opendbc/safety/modes/rivian.h"
+#include "opendbc/safety/modes/subaru.h"
+#include "opendbc/safety/modes/subaru_preglobal.h"
+#include "opendbc/safety/modes/mazda.h"
+#include "opendbc/safety/modes/nissan.h"
+#include "opendbc/safety/modes/volkswagen_mqb.h"
+#include "opendbc/safety/modes/volkswagen_pq.h"
+#include "opendbc/safety/modes/elm327.h"
+#include "opendbc/safety/modes/body.h"
 
 // CAN-FD only safety modes
 #ifdef CANFD
-#include "opendbc/safety/safety/safety_hyundai_canfd.h"
+#include "opendbc/safety/modes/hyundai_canfd.h"
 #endif
 
 uint32_t GET_BYTES(const CANPacket_t *msg, int start, int len) {
