+++ opendbc/safety/tests/misra/test_mutation.py
@@ -19,7 +19,7 @@ mutations = [
   # default
   (None, None, False),
   # general safety
-  ("opendbc/safety/safety/safety_toyota.h", "s/if (addr == 0x260) {/if (addr == 1 || addr == 2) {/g", True),
+  ("opendbc/safety/modes/toyota.h", "s/if (addr == 0x260) {/if (addr == 1 || addr == 2) {/g", True),
 ]
 
 patterns = [
